add unit tests for cron job
